rename root files
