+++ _templates/component/new/components_ComponentName_react_rollup.config.mjs.ejs.t
@@ -1,7 +1,7 @@
 ---
 to: components/<%= h.inflection.camelize(name, false) %>/react/rollup.config.mjs
 ---
-import rootRollupConfig from '../../root.rollup.config.mjs'
+import rootRollupConfig from '../../react.rollup.config.mjs'
 import pkg from './package.json' assert { type: 'json' }
 
 export default rootRollupConfig({
