+++ _templates/component/new/components_ComponentName_vue_vite.config.ts.ejs.t
@@ -2,7 +2,7 @@
 to: components/<%= h.inflection.camelize(name, false) %>/vue/vite.config.ts
 ---
 import * as path from 'path'
-import generateViteConfig from '../../root.vite.config'
+import generateViteConfig from '../../vue.vite.config'
 
 export default generateViteConfig({
   entry: path.resolve(__dirname, './index.ts'),
