+++ components/Button/react/rollup.config.mjs
@@ -1,3 +1,3 @@
-import rootRollupConfig from '../../root.rollup.config.mjs'
+import rootRollupConfig from '../../react.rollup.config.mjs'
 
 export default rootRollupConfig({ input: './index.ts' })
