+++ components/Menu/vue/vite.config.ts
@@ -1,5 +1,5 @@
 import * as path from 'path'
-import generateViteConfig from '../../root.vite.config'
+import generateViteConfig from '../../vue.vite.config'
 
 export default generateViteConfig({
   entry: path.resolve(__dirname, './index.ts'),
