+++ components/Modal/react/rollup.config.mjs
@@ -1,4 +1,4 @@
-import rootRollupConfig from '../../root.rollup.config.mjs'
+import rootRollupConfig from '../../react.rollup.config.mjs'
 import pkg from './package.json' assert { type: 'json' }
 
 export default rootRollupConfig({
