+++ components/Spinner/react/rollup.config.mjs
@@ -1,5 +1,5 @@
 import postcss from 'rollup-plugin-postcss'
-import rootRollupConfig from '../../root.rollup.config.mjs'
+import rootRollupConfig from '../../react.rollup.config.mjs'
 import pkg from './package.json' assert { type: 'json' }
 
 export default rootRollupConfig({
